Changed folder structure for seperate GET and POST methods, added controller folder for containing the same. Seperated the connection file to different folder inside model,added comments for code readibilitty
