--- README.md
@@ -1,3 +1,0 @@
-
-# Elfinsurl
-> elfinsurl is a free and open source URL shortener microservice built with javascript
