+++ models/urlSchema.js
@@ -2,7 +2,11 @@ const mongoose = require("mongoose");
 
 
 var Schema = mongoose.Schema;
+/*Schema for URL,
 
+| original_url(String) | short_url(String) |
+
+*/
 const url_schema = new Schema({
   original_url : String,
   short_url  : String
